Move ProgressBar to watch module
